backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-jose>=3.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
